module config editor adjustment,nav
